+++ example/lib/main_perfect.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_credit_card/src/credit_card_widget_perfect.dart';
-import 'package:flutter_credit_card/src/credit_card_form_perfect.dart';
-import 'package:flutter_credit_card/src/credit_card_model.dart';
+import 'package:animated_payment_card/src/credit_card_widget_perfect.dart';
+import 'package:animated_payment_card/src/credit_card_form_perfect.dart';
+import 'package:animated_payment_card/src/credit_card_model.dart';
 
 void main() {
   runApp(const MyApp());